requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
Pillow==10.2.0
gunicorn==21.2.0
